+++ src/main/java/dev/muon/medieval/client/MedievalClient.java
@@ -1,0 +1,16 @@
+package dev.muon.medieval.client;
+
+import dev.muon.medieval.Medieval;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.client.event.RegisterGuiOverlaysEvent;
+import net.minecraftforge.client.gui.overlay.VanillaGuiOverlay;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.common.Mod;
+
+@Mod.EventBusSubscriber(modid = Medieval.MODID, value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
+public class MedievalClient {
+    @SubscribeEvent
+    public static void onRegisterOverlays(RegisterGuiOverlaysEvent event) {
+        event.registerAbove(VanillaGuiOverlay.PLAYER_HEALTH.id(), "medieval_bars", MedievalGuiOverlays.MEDIEVAL_BARS);
+    }
+}